Reverting Go package name changes. (#1963)
* Reverting Go package name changes.

There were a handful of packages where the names varied from what we
previously shipped. There were a vareity of reasons for that, from
matching the RP instead of the service name, to just straight more
readable ones. This commit returns ensures that there not differing
specifiers to reduce the amount of effort that it takes to adopt using
the new Go packages for folks coming from the previously generated
packages.

* Updating to just modify `redis`