[Azure Search] Moving 2016-09-01-preview spec from private repo (#2427)
* [Azure Search] Moving 2016-09-01-preview spec from private repo

Prior to the reorganization of the azure-rest-api-specs repo structure, the
Swagger specs for the Azure Search data plane API version 2016-09-01-preview
lived in the search-service branch of the azure-rest-api-specs-pr repo. Now
that there is a place for preview/incomplete Swagger specs in the public repo,
we'd like to move this spec there so all the specs are in one place.

The specs themselves have not changed; This is just a repo-to-repo copy.

Like all Azure Search data plane specs, these are incomplete. See this issue
for details: https://github.com/Azure/azure-rest-api-specs/issues/1085

* [Azure Search] Adding 2016-09-01-preview config to readme.md