Truncate the output block if the text is too long
GitHub has a limit of 64k on the text field, so if there are more
results than that we can't show them. Instead we now display a message
directing the user to the build log.

We also produce better output formatting for the build log since people
may need to refer to it now.