Update UserAssignedIdentities spec in Web (#5288)
* Update UserAssignedIdentities spec in Web

Microsoft.Web recently added support for UserAssigned managed service identities (in preview), but the swagger spec is incorrect for the identity payload as it follows the old construct of sending an array of identityIds as opposed to the updated spec which instead sends a dictionary object called userAssignedIdentities. 

Spec + history outlined here:
http://sharepoint/sites/AzureUX/Sparta/_layouts/15/WopiFrame2.aspx?sourcedoc=%7b633632B4-02E4-4DE2-92A9-EDCAF9BE73D4%7d&file=ARM%2BMSI.docx&action=default 

Our swagger contains identityids[] because of previous work to future support UserAssigned identities, but never released the feature in that state, rather we opted to wait for the new spec to be supported and then integrated with that on our end. 

Our MSI implementation reflects Microsoft.Compute's spec for user assigned.

https://github.com/Azure/azure-rest-api-specs/blob/78e1feab142c605839aa3cec23442ca26b8ef04d/specification/compute/resource-manager/Microsoft.Compute/stable/2018-06-01/compute.json

* fix linterdiff tool