Fix: Capacity is a mandatory Property (#8686)
* capacity is a mandatory parameter

* fix examples and model validation issue

* fix prettier check

There are a number of failures in Swagger checks. For breaking change check, given this is preview version and breaking change is intentional, it is accepted. The SDK automation check failures seems unrelated to the code change, with exception of .NET SDK where breaking change did cause build failure. You would need to correct the test code in your PR to csharp SDK repo.